Frontend for Login, migrations, routing.
Frontend for Login, Signup, Password Recovery and the start of the navbar and the main page, along with database migrations, backend file structure and general routing for backend and frontend. Nginx reverse proxy set up too, although some tinkering is required for this. Most of the signup functionality is done but also requires tinkering